rad-publish: Initialize
Calls `git push rad`, then `rad sync`.
